Moves events and news to search-index under one endpoint.
