+++ src/lib/components/ContentfulRichText/nodes/EntryHyperlink.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type { Node as NodeType, EntryHyperlink } from "@contentful/rich-text-types";
-  import type { PageMetadataMapItem } from "$lib/loadPageMetadataMap";
+  import type { MetadataMapItem } from "$lib/loadPageMetadataMap";
   import Node from "./Node.svelte";
   import { getContext } from "svelte";
   import { linksKey, type LinksContext } from "../context";
@@ -21,7 +21,7 @@
 
   const entry = linksContext.linksEntriesMaps.hyperlink.get(entryID);
 
-  let entryMetadata: PageMetadataMapItem;
+  let entryMetadata: MetadataMapItem;
   if (entry?.__typename === "PageMetadata") {
     linksContext.pageMetadataMap.get(entryID);
     const meta = linksContext.pageMetadataMap.get(entryID);
