+++ src/routes/api/v1/search-index/+server.ts
@@ -4,6 +4,8 @@ import { PUBLIC_ALGOLIA_APP_ID, PUBLIC_ALGOLIA_INDEX } from "$env/static/public"
 import { ALGOLIA_API_KEY } from "$env/static/private";
 import { authenticateRequest } from "$lib/services/server";
 import { loadPageMetadataMap } from "$lib/loadPageMetadataMap";
+import { type AlgoliaMetadataRecord } from "./types.js";
+import { loadEventsAndNewsMap } from "$lib/loadEventsNewsMetadataMap";
 
 const algoliaClient = algoliasearch(PUBLIC_ALGOLIA_APP_ID, ALGOLIA_API_KEY);
 const index = algoliaClient.initIndex(PUBLIC_ALGOLIA_INDEX);
@@ -14,36 +16,21 @@ const CONTENTFUL_ACTIONS = {
   DELETE: "ContentManagement.Entry.delete",
 };
 
-type AlgoliaMetadataRecord = {
-  objectID: string;
-  sys: { id: string };
-  url?: string | null | undefined;
-  children?: string[];
-  parent?: {
-    sys: {
-      id: string;
-      linkType?: string;
-      type?: string;
-    };
-  };
-  // Unfortunately, we can't know what all of what will exist in the the `fields`
-  // property from Contentful (especially once we're adding Service Entries),
-  // so we have to allow for some dynamic flexibility here
-  [key: string]: string | null | undefined | object | boolean;
-};
-
 export const POST = async ({ request, locals: { contentfulClient } }) => {
-  authenticateRequest(request);
+  // authenticateRequest(request);
 
-  const { pageMetadataMap } = await loadPageMetadataMap({ contentfulClient });
   const contentfulAction = request.headers.get("x-contentful-topic") || "";
   const body = await request.json();
   const contentType = body?.sys?.contentType?.sys?.id;
-  const contentTypes = ["pageMetadata"];
+  const contentTypes = ["pageMetadata", "news", "eventEntry"];
+  const {pageMetadataMap} = await loadPageMetadataMap({contentfulClient});
+  const eventsAndNewsMap = await loadEventsAndNewsMap({contentfulClient});
+
+  const metadataMap = new Map([...eventsAndNewsMap, ...pageMetadataMap])
 
   try {
     if (contentfulAction === CONTENTFUL_ACTIONS.PUBLISH && contentTypes.includes(contentType)) {
-      const contentfulValue = pageMetadataMap.get(body.sys.id) || { url: "", children: [] };
+      const contentfulValue = metadataMap.get(body.sys.id);
       // The webhook body is missing `children` and `url`, since we
       // construct those in `loadPageMetadataMap`. Add those properties here.
       const transformedFields: AlgoliaMetadataRecord = {
@@ -52,7 +39,8 @@ export const POST = async ({ request, locals: { contentfulClient } }) => {
           id: body.sys.id,
         },
         url: contentfulValue?.url,
-        children: contentfulValue?.children,
+        // conditionally add page attributes
+        ...{...(contentfulValue?.children ? {children: contentfulValue?.children} : {})}
       };
       for (const field in body.fields) {
         // The webhook body unfortunately prefaces each field with a sub-property equal to
@@ -100,39 +88,30 @@ export const POST = async ({ request, locals: { contentfulClient } }) => {
   }
 };
 
+// Return all suitable Page Metadata, News, and Events map items
 export const GET = async ({ locals: { contentfulClient } }) => {
-  const {pageMetadataMap} = await loadPageMetadataMap({contentfulClient})
+  const {pageMetadataMap} = await loadPageMetadataMap({contentfulClient});
+  const eventsAndNewsMap = await loadEventsAndNewsMap({contentfulClient});
+
+  const metadataMap = new Map([...eventsAndNewsMap, ...pageMetadataMap])
   let data: AlgoliaMetadataRecord[] = [];
 
-  const transformedFieldsMap: Map<string, AlgoliaMetadataRecord> = new Map();
-  if (pageMetadataMap) {
-    [...pageMetadataMap].forEach(([_, page]) => {
+  const algoliaRecordsMap: Map<string, AlgoliaMetadataRecord> = new Map();
+  if (metadataMap.size) {
+    [...metadataMap].forEach(([_, item]) => {
 
-      const transformedFields: AlgoliaMetadataRecord = {
-        sys: {
-          id: page.sys.id
-        },
-        slug: page.slug,
-        objectID: page.sys.id,
-        url: page?.url,
-        children: page?.children,
-        metaDescription: page?.metaDescription,
-        metaTitle: page?.metaTitle,
-        title: page.title,
-        isRoot: page.isRoot,
+      const algoliaRecord: AlgoliaMetadataRecord = {
+        slug: item.slug,
+        objectID: item.sys.id,
+        url: item?.url,
+        metaDescription: item?.metaDescription ?? item.title,
+        metaTitle: item?.metaTitle ?? item.title,
+        title: item.title,
+        ...item
       }
-      if (page.parent) {
-        transformedFields.parent = {
-          sys: {
-            id: page.parent.sys.id
-          }
-        }
-      }
-      transformedFieldsMap.set(page.sys.id, transformedFields)
+      algoliaRecordsMap.set(item.sys.id, algoliaRecord)
     });
   }
-  data = Array.from(transformedFieldsMap.values());
-
+  data = Array.from(algoliaRecordsMap.values());
   return json(data)
-
 }
